add round intrinsic

+++ dataset/GitHub_Java/oracle.graal/sulong/projects/com.oracle.truffle.llvm.runtime/src/com/oracle/truffle/llvm/runtime/nodes/intrinsics/c/LLVMCMathsIntrinsics.java
@@ -253,7 +253,7 @@ public abstract class LLVMCMathsIntrinsics {
 
         @Specialization
         protected float doIntrinsic(float value) {
-            return Math.round(value);
+            return (float) Math.round(value);
         }
 
         @Specialization
